Separate the in-memory mocks from the ones that spawn real processes (or at least try to).
